+++ foliplus/__init__.py
@@ -13,6 +13,7 @@ except ImportError:
 from .BaseControl import BaseControl
 from .ExportControl import ExportControl
 from .FullscreenControl import FullscreenControl
+from .GuideControl import GuideControl
 from .HeatmapControl import HeatmapControl
 from .LayerControl import LayerControl
 from .LocateControl import LocateControl
@@ -28,6 +29,7 @@ __all__ = [
     "LayerControl",
     "LocateControl",
     "MeasureControl",
+    "GuideControl",
     "ScaleControl",
     "SearchControl",
 ]